+++ project/map-backend/src/main/java/com/stevenst/app/service/MarkerService.java
@@ -12,5 +12,5 @@ public interface MarkerService {
 	
 	Marker addMarker(Marker marker);
 
-	Marker getMarker(int id);
+	Marker getMarker(Long id);
 }
